requirements.txt
numpy
pandas 
matplotlib
torch
polars-lts-cpu
scikit-learn
seaborn